fix bug eddying conference and presentation

+++ src/main/java/ua/krasun/conference_portal_servlet/model/dao/GenericDao.java
@@ -8,7 +8,7 @@ public interface GenericDao<T> extends AutoCloseable {
 
     void add(T entity) throws SQLException;
 
-    T findById(int id);
+    T findById(long id);
 
     List<T> findAll();
